requirements.txt
langchain
langchain-cohere
langchain-community
langchain-text-splitters
langchain-core
chromadb
pypdf
python-dotenv
streamlit
cohere
duckduckgo-search